+++ public/changelog.md
@@ -1,2 +1,21 @@
+### Coming Soonish
+- Add Quick Log feature to easily log everything that has happened in a single post
+- Get Google Drive syncing working in an easy-to-use and unobtrusive way
+- Less hard coding, more flexible templates
+  - evolution chain as template instead of hard coding
+  - list item templates - Ex: "Pokemon.levelLogs.item"
+  - better link wrapping/inclusion - Ex: {{field_link}}
+- Allow custom fields for Pokémon profiles (straight replacement only) to allow for extra customization
+- Autofill Pokemon data from FizzyDex data
 
-AAAAAAA
+### v0.1.1
+- Feature: Home Page - Improved Change Log
+- Fix: Pokémon Page - Evolution Stage 3 Method Link input changing wrong field
+- Fix: Settings Page - Textarea does not have line breaks on Firefox
+- Fix: Bond Page - Output dates appearing as "+0000"
+- Fix: Spreadsheet Import loads "Gender" into "Ability" 
+
+### v0.1 - Initial Release
+- Pages: Trainers, Pokemon, Items, Wallet, Bond, Word Counter, Settings, and a hidden-by-default Data page.
+- Feature: Import from spreadsheet downloaded as a copy of the previous tracker that was based in Google Sheets.
+- Feature: Import (Upload) from and Export (Download) to file on your local device.